comments

+++ diana/access/decompress.py
@@ -8,32 +8,37 @@ import zipfile
 from typing import Optional
 
 
-def decompress_gzip_file(compressed_file_name: str, size: int = 1048576) -> str:
+def decompress_gzip_file(compressed_file_name: str, size: int = 1048576, remove: bool = True) -> str:
     """
     Decompresses a gzip compressed file and removes the compressed file.
 
     Args:
         compressed_file_name: The file name of the compressed file.
         size: The buffer size to process the decompression at.
+        remove: Whether to remove the compressed file following decompression.
 
     Returns:
         The file name of the decompressed file.
     """
+    # Determine the file name of the decompressed file.
     decompressed_file_name = os.path.splitext(compressed_file_name)[0]
 
+    # Decompress the file.
     if not os.path.exists(decompressed_file_name):
         with open(decompressed_file_name, "wb") as decompressed_file:
             with gzip.open(compressed_file_name, "rb") as compressed_file:
                 while chunk := compressed_file.read(size):
                     decompressed_file.write(chunk)
 
-    os.remove(compressed_file_name)
+    # Remove the compressed file if applicable.
+    if remove:
+        os.remove(compressed_file_name)
     return decompressed_file_name
 
 
 def decompress_zip_file(
         compressed_file_name: str,
-        file_from_zip_archive: Optional[re.Pattern[str]] = None) -> str:
+        file_from_zip_archive: Optional[re.Pattern[str]] = None, remove: bool = True) -> str:
     """
     Decompresses a zip compressed file and removes the compressed file.
 
@@ -42,10 +47,12 @@ def decompress_zip_file(
         file_from_zip_archive: The file from the zip archive to extract. If
             this file is not specified or found, the first in the archive is
             used. The first file matching the regular expression is used.
+        remove: Whether to remove the compressed file following decompression.
 
     Returns:
         The file name of the decompressed file.
     """
+    # Determine the file name of the decompressed file and decompress the file.
     with zipfile.ZipFile(compressed_file_name) as archive:
         if not file_from_zip_archive:
             file = archive.namelist()[0]
@@ -74,5 +81,7 @@ def decompress_zip_file(
                 f"{os.path.splitext(os.path.basename(sys.argv[0]))[0]}-"
                 f"{os.getpid()}", file)
 
-    os.remove(compressed_file_name)
+    # Remove the compressed file if applicable.
+    if remove:
+        os.remove(compressed_file_name)
     return decompressed_file_name
